+++ migration-newsapi-to-ts/src/components/view/appView.ts
@@ -1,17 +1,7 @@
 import News from "./news/news";
 import Sources from "./sources/sources";
-import { NewsData, SourcesData } from "../../types";
-
-interface IRenderNews {
-  drawNews(data: NewsData): void;
-  drawSources(data: SourcesData): void;
-}
-
-// Viewer
-interface IViewer extends IRenderNews {
-  readonly news: News;
-  readonly sources: Sources;
-}
+import { NewsData, SourcesData } from "../../types/types";
+import { IViewer } from "../../types/interfaces";
 
 // Class
 export class AppView implements IViewer {
@@ -23,12 +13,12 @@ export class AppView implements IViewer {
     this.sources = new Sources();
   }
 
-  drawNews(data: NewsData): void {
+  public drawNews(data: NewsData): void {
     const values: NewsData["articles"] = data?.articles ? data.articles : [];
     this.news.draw(values);
   }
 
-  drawSources(data: SourcesData): void {
+  public drawSources(data: SourcesData): void {
     const values: SourcesData["sources"] = data?.sources ? data.sources : [];
     this.sources.draw(values);
   }
